Nirvaan/backend-old/requirements.txt
fastapi
uvicorn
python-dotenv
retell-sdk
pydantic